requirements.txt
cffi==1.7.0
Django==1.10
django-classy-tags==0.7.2
django-registration==2.1.2
django-simple-captcha==0.5.2
django-taggit==0.20.2
django-taggit-templatetags2==1.5.2
fake-factory==0.6.0
houdini.py==0.1.0
ipaddress==1.0.16
misaka==2.0.0
MySQL-python==1.2.5
Pillow==3.3.1
psutil==4.3.0
pycparser==2.14
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
qiniu==7.0.8
requests==2.11.1
six==1.10.0
